requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docopt==0.6.2
facenet-pytorch==2.5.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
Pillow==8.1.2
pipreqs==0.4.10
plotly==4.14.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
yarg==0.1.9
